pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "anthropic>=0.64.0",
    "bcrypt>=4.3.0",
    "cryptography>=45.0.6",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-dance>=7.1.0",
    "flask-jwt-extended>=4.7.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "httpx>=0.28.1",
    "notion-client>=2.4.0",
    "oauthlib>=3.3.1",
    "openai>=1.99.9",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "slack-sdk>=3.36.0",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "tenacity>=9.1.2",
    "pinecone-client>=6.0.0",
    "celery>=5.5.3",
    "apscheduler>=3.11.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.30",
    "langchain-anthropic>=0.3.18",
    "langchain-community>=0.3.27",
    "pinecone>=7.3.0",
]
